Add task watchdog
This makes sure command processing is making progress.
If for whatever reason we do not get a command for 5
minutes, reboot.  This catches the other random hangs in
the app which are now quite rare, but better than actually
having to hardware reset the device, especially if it is
connected to a battery in a case.